cleaner with pythonpaths
