+++ compiler/src/main/scala/edg/compiler/PythonInterface.scala
@@ -21,12 +21,20 @@ object ProtobufStdioSubprocess {
 }
 
 
-class ProtobufStdioSubprocess
-    [RequestType <: scalapb.GeneratedMessage, ResponseType <: scalapb.GeneratedMessage](
+class ProtobufStdioSubprocess[RequestType <: scalapb.GeneratedMessage, ResponseType <: scalapb.GeneratedMessage](
     responseType: scalapb.GeneratedMessageCompanion[ResponseType],
-    args: Seq[String]) {
+    pythonPaths: Seq[String], args: Seq[String]) {
   protected val process: Either[Process, Throwable] = try {
-    Left(new ProcessBuilder(args: _*).start())
+    val processBuilder = new ProcessBuilder(args: _*)
+    if (pythonPaths.nonEmpty) {
+      val env = processBuilder.environment()
+      val pythonPathString = pythonPaths.mkString(";")
+      Option(env.get("PYTHONPATH")) match {  // merge existing PYTHONPATH if exists
+        case None => env.put("PYTHONPATH", pythonPathString)
+        case Some(envPythonPath) => env.put("PYTHONPATH", envPythonPath + ";" + pythonPathString)
+      }
+    }
+    Left(processBuilder.start())
   } catch {
     case e: Throwable => Right(e)  // if it fails store the exception to be thrown when we can 
   }
@@ -114,7 +122,7 @@ class ProtobufStdioSubprocess
 
 
 object PythonInterface {
-  private val kHdlServerFilePath = "HdlInterfaceServer.py"
+  private val kHdlServerFilePath = "edg_hdl_server/__main__.py"
   // returns the HDL server Python script if it exists locally, otherwise returns None.
   def serverFileOption(root: Option[File] = None): Option[File] = {
     val hdlServerFile = root match {
@@ -136,13 +144,13 @@ object PythonInterface {
   *
   * If the serverFile is specified, run that; otherwise use "python -m edg_hdl_server" for the global package.
   */
-class PythonInterface(serverFile: Option[File], pythonInterpreter: String = "python") {
+class PythonInterface(serverFile: Option[File], pythonPaths: Seq[String], pythonInterpreter: String = "python") {
   val command = serverFile match {  // -u for unbuffered mode
     case Some(serverFile) => Seq(pythonInterpreter, "-u", serverFile.getAbsolutePath)
     case None => Seq(pythonInterpreter, "-u", "-m", "edg_hdl_server")
   }
   protected val process = new ProtobufStdioSubprocess[edgrpc.HdlRequest, edgrpc.HdlResponse](
-    edgrpc.HdlResponse, command)
+    edgrpc.HdlResponse, pythonPaths, command)
   val processOutputStream: InputStream = process.outputStream
   val processErrorStream: InputStream = process.errorStream
