+++ compiler/src/test/scala/edg/compiler/PythonInterfaceTest.scala
@@ -13,7 +13,8 @@ class PythonInterfaceTest extends AnyFlatSpec {
     val compiledDir = new File(getClass.getResource("").getPath)
     // above returns compiler/target/scala-2.xx/test-classes/edg/compiler, get the root repo dir
     val repoDir = compiledDir.getParentFile.getParentFile.getParentFile.getParentFile.getParentFile.getParentFile
-    val pyIf = new PythonInterface(Some(new File(repoDir, "HdlInterfaceServer.py")))
+    val pyIf = new PythonInterface(Some(new File(repoDir, "edg_hdl_server/__main__.py")),
+      Seq(new File(".").getAbsolutePath))
     pyIf.indexModule("edg_core").getClass should equal(classOf[Errorable.Success[Seq[LibraryPath]]])
     pyIf.shutdown() should equal(0)
   }
