--- edg_hdl_server/__init__.py
@@ -1,1 +1,0 @@
-from .Server import run_server
